Mise en place page role/admin

+++ client/src/views/admin/rôles/Roles.jsx
@@ -5,32 +5,50 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faPenToSquare } from '@fortawesome/free-regular-svg-icons';
 import { faTrash } from "@fortawesome/free-solid-svg-icons";
 
+import ValidateModal from "../../../components/modal/ValidateModal";
 import "./roles.scss";
 
 function Roles() {
+  document.title = "Back Office || Roles";
   const [roles, setRoles] = useState(null);
   const [newRoleLabel, setNewRoleLabel] = useState("");
   const [shouldRefreshRoles, setShouldRefreshRoles] = useState(false);
   const [editingRoleId, setEditingRoleId] = useState(null);
   const [editingRoleLabel, setEditingRoleLabel] = useState("");
+  const [error, setError] = useState(null);
+  const [success, setSuccess] = useState(false);
+  const [showModal, setShowModal] = useState(false);
+  const [roleToDelete, setRoleToDelete] = useState(null);
 
   useEffect(() => {
-    document.title = "Back Office || Roles";
     async function fetchRoles() {
-      const response = await fetch(
-        "http://localhost:9000/api/v1/roles",
-        {
-          credentials: "include",
+      try {
+        const response = await fetch(
+          "http://localhost:9000/api/v1/roles",
+          {
+            method: "GET",
+            headers: {
+              "Accept" : "application/json",
+            },
+            credentials: "include",
+          }
+        );
+
+        if (!response) {
+          setError("Aucun rôles trouvés");
+          setSuccess(false);
+          return;
         }
-      );
 
-      if (!response) {
-        return;
-      }
-
-      if (response.ok) {
-        const data = await response.json();
-        setRoles(data.response);
+        if (response.ok) {
+          const data = await response.json();
+          setRoles(data.response);
+          setError(null);
+          setSuccess(false);
+        }
+      } catch (error) {
+        setError("Erreur de réseau");
+        setSuccess(null);
       }
     }
     fetchRoles();
@@ -55,11 +73,15 @@ function Roles() {
       if (response.ok) {
         setNewRoleLabel(""),
         setShouldRefreshRoles(prev => !prev)
+        setError(null);
+        setSuccess(true);
       } else {
-        console.error("Erreur lors de l'ajout du rôle");
+        setError("Erreur lors de l'ajout du rôle");
+        setSuccess(false);
       }
     } catch (error) {
-      console.log("Erreur:", error);
+      setError("Erreur réseaux", error);
+      setSuccess(false);
     }
   }
 
@@ -83,11 +105,15 @@ function Roles() {
         setEditingRoleId(null);
         setEditingRoleLabel("");
         setShouldRefreshRoles(prev => !prev);
+        setError(null);
+        setSuccess(false);
       } else {
-        console.log("Erreur lors de la modification du rôle");
+        setError("Erreur lors de la modification du rôle");
+        setSuccess(false);
       }
     } catch (error) {
-      console.log("Erreur:", error);
+      setError("Erreur reseaux", error);
+      setSuccess(false);
     }
   }
 
@@ -96,12 +122,21 @@ function Roles() {
     setEditingRoleLabel(role.label);
   }
 
-  // Suppresion de roles
-  async function handleDelete(e, id) {
+  // Affichage de la modal
+  function handleDeleteClick(e, role) {
     e.preventDefault();
+    e.stopPropagation();
+    setRoleToDelete(role);
+    setShowModal(true);
+  }
+
+  // Suppresion de roles
+  async function handleValidateDelete() {
+    if (!roleToDelete) return;
+
     try {
       const response = await fetch(
-        `http://localhost:9000/api/v1/roles/${id}`,
+        `http://localhost:9000/api/v1/roles/${roleToDelete.id}`,
         {
           method: "DELETE",
           credentials: "include",
@@ -110,13 +145,26 @@ function Roles() {
 
       if (response.ok) {
         setShouldRefreshRoles(prev => !prev);
+        setError(null);
+        setSuccess(true);
       } else {
-        console.log("Erreur lors de la suppression du role");
+        setError("Erreur lors de la suppression du role");
+        setSuccess(false);
       }
     } catch (error) {
-      console.log("Erreur:", error)
+      setError("Erreur reseaux", error)
+      setSuccess(false);
+    } finally {
+      setShowModal(false);
+      setRoleToDelete(null);
     }
   }
+
+    // Annuler la suppression
+    function handleCancelDelete() {
+      setShowModal(false);
+      setRoleToDelete(null);
+    }
 
   if (!roles) {
     return (
@@ -135,10 +183,20 @@ function Roles() {
   }
 
   return (
-    <main>
-      <h1>Bienvenue sur la page Roles</h1>
-      <section>
-        {/* Table roles */}
+    <main id="gestion-role">
+      <h1>Gestion des rôles</h1>
+
+      {/* Affichage de l'erreur */}
+      {error && <div className="error-message">{error}</div>}
+
+      {/* Affichage du message de succès */}
+      { success && (
+        <div className="success-message">
+          Le rôle à été ajouté avec succès !
+        </div>
+      )}
+
+      <section className="role-table">
         <table>
           <thead>
             <tr>
@@ -153,12 +211,11 @@ function Roles() {
                 <tr key={role.id}>
                   <td>{role.id}</td>
                   <td>{role.label}</td>
-                  {/* Actions */}
                   <td>
-                    <button onClick={() => handleEditClick(role)}>
+                    <button className="btn-edit" onClick={() => handleEditClick(role)}>
                       <FontAwesomeIcon icon={faPenToSquare} />
                     </button>
-                    <button onClick={(e) => handleDelete(e, role.id)}>
+                    <button className="btn-delete" onClick={(e) => handleDeleteClick(e, role)}>
                       <FontAwesomeIcon icon={faTrash} />
                     </button>
                   </td>
@@ -168,33 +225,36 @@ function Roles() {
           </tbody>
         </table>
 
-        <h3>Ajouter une nouvelle categorie</h3>
-        <form onSubmit={handleAddRole}>
-            <label>
-              Nom du role
+        <section className="add-edit-role">
+        <h2>Ajouter un rôle</h2>
+        <form className="form-role" onSubmit={handleAddRole}>
+            <label htmlFor="role">Nom</label>
               <input
                 type="text"
+                id="role"
+                placeholder="Nom du rôle"
+                aria-label="Nom du rôle"
                 value={newRoleLabel}
                 onChange={(e) => setNewRoleLabel(e.target.value)}
                 required
               />
-            </label>
             <button type="submit">Ajouter</button>
         </form>
+        </section>
 
+        <section className="add-edit-role">
         {editingRoleId && (
           <>
-            <h3>Modifier une catégorie</h3>
-            <form onSubmit={handleEditRole}>
-              <label>
-                Nom de la catégorie
+            <h2>Modifier un rôle</h2>
+            <form className="form-role" onSubmit={handleEditRole}>
+              <label htmlFor="role-edit">Nom</label>
                 <input
                   type="text"
+                  id="role-edit"
                   value={editingRoleLabel}
                   onChange={(e) => setEditingRoleLabel(e.target.value)}
                   required
                 />
-              </label>
               <button type="submit">Modifier</button>
               <button type="button" onClick={() => setEditingRoleId(null)}>
                 Annuler
@@ -202,8 +262,17 @@ function Roles() {
             </form>
           </>
         )}
+        </section>
         <Link to={"/admin/dashboard"}>Retour au tableau de bord</Link>
       </section>
+
+      {/* Modal de validation */}
+      <ValidateModal
+        show={showModal}
+        message="Êtes-vous sûr de vouloir supprimer ce role ?"
+        onConfirm={handleValidateDelete}
+        onCancel={handleCancelDelete}
+      />
     </main>
   );
 }
